semana06/dia6/requirements.txt
Django==3.0.5
djangorestframework==3.13.1
django-rest-framework-mongoengine==3.4.1
mongoengine==0.24.1
pymongo==3.12.3
